+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 # Change Log
 
 All notable changes to the "copy-with-context" extension will be documented in this file.
+
+## [1.5.0] - 05-06-2025
+
+### Added
+- Interactive UI for editing the excluded files/folders, use the=is feature from command pallete (shift+cmd+p) and look for "Copy Context: Edit Excluded Paths", this will open the quick pick dialogue to add/remove files
+- Performance optimizations with caching of excluded paths